requirements.txt
fastapi==0.110.0
uvicorn==0.20.0
sqlalchemy[asyncio]==2.0.0
asyncpg==0.29.0
pydantic-settings==2.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
greenlet==3.0.3 
bcrypt==4.2.0

pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-mock==3.14.0
aiosqlite==0.21.0
dnspython==2.7.0
email-validator==2.2.0